requirements.txt
# Framework principal
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Base de datos y ORM
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0

# Validación y configuración
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1

# Autenticación y seguridad
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.18

# Utilidades
email-validator==2.2.0
